[refactor] entry point(app.js) 추상화를 위해 폴더 구조 변경
